requirements.txt
streamlit>=1.20.0
requests>=2.28.0
beautifulsoup4>=4.12.0
google-genai>=0.13.0
tqdm>=4.65.0
python-dotenv>=1.0.0
feedparser>=6.0.10
pandas>=2.1.0
matplotlib>=3.7.0
